requirements.txt
pyodbc==4.0.22
flake8==3.5.0
yamllint==1.8.1
pyyaml==3.12
